requirements.txt
absl-py==2.1.0
addict==2.4.0
ansitable==0.11.1
asttokens==2.4.1
attrs==23.2.0
blinker==1.8.2
certifi==2024.7.4
cfgv==3.4.0
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
colored==2.2.4
comm==0.2.2
ConfigArgParse==1.7
contourpy==1.2.1
cycler==0.12.1
dash==2.17.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
decorator==5.1.1
distlib==0.3.8
docker-pycreds==0.4.0
etils==1.5.2
evdev==1.7.1
exceptiongroup==1.2.2
executing==2.0.1
Farama-Notifications==0.0.4
fastjsonschema==2.20.0
filelock==3.15.4
Flask==3.0.3
fonttools==4.53.1
fsspec==2024.6.1
gitdb==4.0.11
GitPython==3.1.43
glfw==2.7.0
grpcio==1.65.2
gymnasium==0.29.1
h5py==3.11.0
hidapi==0.14.0.post2
huggingface-hub==0.24.5
identify==2.6.0
idna==3.7
imageio==2.34.2
importlib_metadata==8.2.0
importlib_resources==6.4.0
ipython==8.18.1
ipywidgets==8.1.3
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_core==5.7.2
jupyterlab_widgets==3.0.11
kiwisolver==1.4.5
Klampt==0.9.1.post5
lightning==2.5.0.post0
llvmlite==0.43.0
Markdown==3.6
MarkupSafe==2.1.5
matplotlib==3.9.1
matplotlib-inline==0.1.7
mpmath==1.3.0
mujoco==3.2.0
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.2.1
nodeenv==1.9.1
numba==0.60.0
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.20
nvidia-nvtx-cu12==12.1.105
nvisii==1.1.72
open3d==0.18.0
opencv-python==4.10.0.84
packaging==24.1
pandas==2.2.2
parso==0.8.4
pexpect==4.9.0
pgraph-python==0.6.2
pillow==10.4.0
platformdirs==4.2.2
plotly==5.23.0
pre-commit==3.8.0
progress==1.6
prompt_toolkit==3.0.47
protobuf==4.25.4
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pybullet-svl==3.1.6.4
Pygments==2.18.0
pynput==1.7.7
PyOpenGL==3.1.7
pyparsing==3.1.2
pyquaternion==0.9.9
python-dateutil==2.9.0.post0
python-xlib==0.33
pytz==2024.1
PyYAML==6.0.1
redis==5.2.1
referencing==0.35.1
requests==2.32.3
retrying==1.3.4
#robosuite==1.5.1
roboticstoolbox-python==1.1.1
rpds-py==0.19.1
rtb-data==1.0.1
sb3_contrib==2.3.0
scikit-image==0.24.0
scikit-learn==1.5.1
scipy==1.13.1
sentry-sdk==2.12.0
setproctitle==1.3.3
shapely==2.0.6
six==1.16.0
smmap==5.0.1
spatialgeometry==1.1.0
spatialmath-python==1.1.10
stable_baselines3==2.3.2
stable-baselines3[extra]==2.3.2
stack-data==0.6.3
swift-sim==1.1.0
sympy==1.13.1
tabulate==0.9.0
tenacity==9.0.0
tensorboard==2.17.0
tensorboard-data-server==0.7.2
termcolor==2.4.0
threadpoolctl==3.5.0
torch==2.4.0
torchvision==0.19.0
tqdm==4.66.4
traitlets==5.14.3
transforms3d==0.4.2
triton==3.0.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
virtualenv==20.26.3
wandb==0.17.5
wasabi==1.1.3
wcwidth==0.2.13
websockets==12.0
Werkzeug==3.0.3
widgetsnbextension==4.0.11
zipp==3.19.2
